Fixed UpdateEpisode endpoint. Changed IList to List for greater flexibility. Changed long to int for many properties.

+++ SonarrSharp/Models/AlternateTitle.cs
@@ -21,6 +21,6 @@ namespace SonarrSharp.Endpoints.Series.Data
         /// <value>
         /// The season number.
         /// </value>
-        [J("seasonNumber")] public long SeasonNumber { get; set; }
+        [J("seasonNumber")] public int SeasonNumber { get; set; }
     }
 }
